refactor: reorganize package structure
